% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_covs.R
\name{.apply_conversions}
\alias{.apply_conversions}
\title{Apply Conversions to Dataframe}
\usage{
.apply_conversions(df, thresholds)
}
\arguments{
\item{df}{Dataframe. The dataframe containing columns to convert.}

\item{thresholds}{List. A named list of threshold vectors as returned by
\code{\link{.find_thresholds}}.}
}
\value{
The dataframe with specified columns converted to factors.
}
\description{
Converts continuous columns in a dataframe to categorical factors using
the provided thresholds.
}
\keyword{internal}
