% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mat.R
\name{.build_correlation_matrix}
\alias{.build_correlation_matrix}
\title{Build Correlation Matrix}
\usage{
.build_correlation_matrix(correlations, covariates, context)
}
\arguments{
\item{correlations}{List. A list of correlation specification objects, or
\code{NULL}.}

\item{covariates}{List. A list of covariate objects.}

\item{context}{Character string. Either "within" or "between", specifying
which correlation type to use.}
}
\value{
A symmetric correlation matrix with covariate names as dimension
names.
}
\description{
Constructs a correlation matrix from a list of correlation specifications
and covariate objects.
}
\keyword{internal}
