% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mat.R
\name{.build_sd_matrix}
\alias{.build_sd_matrix}
\title{Build Standard Deviation Matrix}
\usage{
.build_sd_matrix(covariates, type)
}
\arguments{
\item{covariates}{List. A list of covariate objects.}

\item{type}{Character string. Either "within" or "between", specifying
which variance component to use.}
}
\value{
A diagonal matrix of standard deviations with covariate names
as dimension names.
}
\description{
Constructs a diagonal matrix of standard deviations from covariate
specifications, using either within-cluster or between-cluster variance.
}
\keyword{internal}
