% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_engine.R
\name{.sim_engine}
\alias{.sim_engine}
\title{Internal engine for clustered variables simulation}
\usage{
.sim_engine(cluster_sizes, mean_vec, D_w, D_b, L_w, L_b)
}
\description{
Assumes:
\itemize{
\item \code{cluster_sizes} is a positive integer vector
\item \code{mean_vec} is a numeric vector of length p
\item \code{sd_vec} is a numeric vector of length p
\item \code{chol_within} and \code{chol_between} are numeric p x p covariance factors
}
}
\details{
Inputs are assumed to be pre-validated by the caller.
}
\keyword{internal}
