% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.unwrap_nested_list}
\alias{.unwrap_nested_list}
\title{Unwrap single nested list}
\usage{
.unwrap_nested_list(x, obj_class)
}
\arguments{
\item{x}{A list to potentially unwrap.}

\item{obj_class}{Character vector of classes to exclude from unwrapping.
Objects inheriting from these classes won't be unwrapped even if they're
the only element in the list.}
}
\value{
The unwrapped list (contents of \code{x[[1]]}) if \code{x} contains exactly one
element that is a list and does not inherit from any class in \code{obj_class}.
Otherwise, returns \code{x} unchanged.
}
\description{
If a list contains exactly one element that is itself a list (but not an
object of a specified class), unwrap it by one level. Otherwise, return the
list unchanged.
}
\examples{
# Unwraps single nested list
.unwrap_nested_list(list(list(a = 1, b = 2)), obj_class = "custom_class")
# Returns: list(a = 1, b = 2)

# Doesn't unwrap if there are multiple elements
.unwrap_nested_list(list(list(a = 1), list(b = 2)), obj_class = "custom_class")
# Returns: list(list(a = 1), list(b = 2))

# Doesn't unwrap objects of the specified class
.unwrap_nested_list(
  list(structure(list(a = 1), class = "ml_covariate")),
  obj_class = "ml_covariate"
)
# Returns: list(structure(list(a = 1), class = "ml_covariate"))

}
\keyword{internal}
