% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_engine.R
\name{.simulation_engine}
\alias{.simulation_engine}
\title{Internal engine for clustered variables simulation}
\usage{
.simulation_engine(cluster_sizes, mean_vec, D_w, D_b, L_w, L_b, seed = NULL)
}
\description{
Internal engine for clustered variables simulation
}
